requirements.txt
streamlit>=1.28.0
tensorflow>=2.16.0
protobuf>=3.20.3
pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.17.0
pandas>=2.0.3
psutil>=5.9.5
